@@ -81,7 +81,11 @@
         }
     },
     "require-dev": {
+        "doctrine/doctrine-fixtures-bundle": "^3.6",
         "phpunit/phpunit": "^11.4",
-        "symfony/maker-bundle": "^1.61"
+        "symfony/browser-kit": "7.1.*",
+        "symfony/http-client": "7.1.*",
+        "symfony/maker-bundle": "^1.61",
+        "zenstruck/foundry": "^2.2"
     }
 }
