@@ -1,5 +1,5 @@
 <?php
-namespace App\Repository;
+namespace App\Infrastructure\Repository;
 
 use Doctrine\Bundle\DoctrineBundle\Repository\ServiceEntityRepository;
 use Doctrine\Persistence\ManagerRegistry;
